Now have spinners for foul strings and the passed in official names. Need to figure out how to allow multi select in those spinners or find some way to be able to select multiple fouls on a given play.
Might make sense to have a second button for adding a foul to a list
and then once the play is saved, use that list then clear it for the next play.

Will need to tie all that information together at some point.

Also need to figure out how I want to save this thing to a file or something
so it can be looked at later.